Revert "Initial Commit (#48)"
This reverts commit fcd043c042a23811f3bec627fc7a8504eda80fd2.
